Update CMakeLists.txt
This entire file is not needed due to the fact that there is not a program reading from this txt file, rather use the functions listed in your __init__ file. I chopped of the ending function because how can we install the ArduinoJson Package if this file does not have access to it in the first place.